Split LauncherFolderItem into separate .cpp & .h files
Gotten confused too many times reading the code with when nearly
everything else is having one bigger class per file.
