+++ src/src.pro
@@ -41,6 +41,7 @@ PUBLICHEADERS += \
     components/launcherwatchermodel.h \
     components/launchermonitor.h \
     components/launcherdbus.h \
+    components/launcherfolderitem.h \
     components/launcherfoldermodel.h \
     notifications/notificationmanager.h \
     notifications/lipsticknotification.h \
@@ -96,6 +97,7 @@ SOURCES += \
     components/launcherwatchermodel.cpp \
     components/launchermonitor.cpp \
     components/launcherdbus.cpp \
+    components/launcherfolderitem.cpp \
     components/launcherfoldermodel.cpp \
     notifications/notificationmanager.cpp \
     notifications/notificationmanageradaptor.cpp \
